# public/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Decent Event Management API",
        "description": "Event ticketing and management system API with registration, payment, ticketing, check-in, and reporting capabilities.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.example.com/v1",
            "description": "Production server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/admin/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "List tickets with optional filters",
                "operationId": "7a17ab85a1ce1e9410f406b9f6af0af0",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by ticket status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ticket_type_id",
                        "in": "query",
                        "description": "Filter by ticket type ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by ticket number or holder name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Results per page, capped at 100",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of tickets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "ulid": {
                                                        "type": "string"
                                                    },
                                                    "ticket_number": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "admits_total": {
                                                        "type": "integer"
                                                    },
                                                    "admitted_count": {
                                                        "type": "integer"
                                                    },
                                                    "price_paid_paisa": {
                                                        "description": "Amount in paisa (1 BDT = 100 paisa)",
                                                        "type": "integer"
                                                    },
                                                    "currency": {
                                                        "type": "string"
                                                    },
                                                    "holder_name": {
                                                        "type": "string"
                                                    },
                                                    "holder_batch_year": {
                                                        "type": "integer",
                                                        "nullable": true
                                                    },
                                                    "holder_type_label": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "issued_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "voided_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "void_reason": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "manifest_version": {
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Missing ticket.view_any permission"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/tickets/{ticket}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Get a single ticket by ULID",
                "operationId": "7c2f803c828792d9eda171bc22209a39",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "Ticket ULID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ulid": {
                                            "type": "string"
                                        },
                                        "ticket_number": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "admits_total": {
                                            "type": "integer"
                                        },
                                        "admitted_count": {
                                            "type": "integer"
                                        },
                                        "price_paid_paisa": {
                                            "description": "Amount in paisa (1 BDT = 100 paisa)",
                                            "type": "integer"
                                        },
                                        "currency": {
                                            "type": "string"
                                        },
                                        "holder_name": {
                                            "type": "string"
                                        },
                                        "holder_batch_year": {
                                            "type": "integer",
                                            "nullable": true
                                        },
                                        "holder_type_label": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "issued_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        },
                                        "voided_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        },
                                        "void_reason": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "first_admitted_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        },
                                        "last_admitted_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        },
                                        "manifest_version": {
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        },
                                        "ticket_type": {
                                            "description": "Ticket type this ticket belongs to (see Ticket Types schema)",
                                            "type": "object"
                                        },
                                        "qr_code_payload": {
                                            "description": "Signed QR payload, when loaded",
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "replaces": {
                                            "description": "The ticket this one replaced via reissue, if any",
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Missing ticket.view permission"
                    },
                    "404": {
                        "description": "Ticket not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token issued by the relevant login endpoint. Admin, attendee, and scanner guards each mint tokens scoped to that guard and cannot be used interchangeably.",
                "bearerFormat": "Sanctum personal access token",
                "scheme": "bearer"
            }
        }
    },
...
}